requirements.txt
pytest
pytest-timeout
pytest-mock
validators
opencv-python
requests
kafka-python
fire